+++ cogs/listing.py
@@ -22,7 +22,7 @@ from typing import TYPE_CHECKING
 
 from discord.ext import commands, tasks
 
-from cogs.utils import Context, list_wrappers
+from src import Context, list_wrappers
 from main import Bot
 
 if TYPE_CHECKING:
